refactor: group bookings hooks in folder
