+++ packages/platform/atoms/index.ts
@@ -9,9 +9,9 @@ export { useEventTypes } from "./hooks/event-types/public/useEventTypes";
 export { useTeamEventTypes } from "./hooks/event-types/public/useTeamEventTypes";
 export { useEventType as useEvent } from "./hooks/event-types/public/useEventType";
 export { useEventTypeById } from "./hooks/event-types/private/useEventTypeById";
-export { useCancelBooking } from "./hooks/useCancelBooking";
-export { useGetBooking } from "./hooks/useGetBooking";
-export { useGetBookings } from "./hooks/useGetBookings";
+export { useCancelBooking } from "./hooks/bookings/useCancelBooking";
+export { useGetBooking } from "./hooks/bookings/useGetBooking";
+export { useGetBookings } from "./hooks/bookings/useGetBookings";
 export { useMe } from "./hooks/useMe";
 export { OutlookConnect } from "./connect/outlook/OutlookConnect";
 export * as Connect from "./connect";
