+++ packages/platform/atoms/monorepo.ts
@@ -3,7 +3,7 @@ export { CalProvider } from "./cal-provider/CalProvider";
 export { useIsPlatform } from "./hooks/useIsPlatform";
 export { useAtomsContext } from "./hooks/useAtomsContext";
 export { useEventTypeById } from "./hooks/event-types/private/useEventTypeById";
-export { useHandleBookEvent } from "./hooks/useHandleBookEvent";
+export { useHandleBookEvent } from "./hooks/bookings/useHandleBookEvent";
 export * as Dialog from "./src/components/ui/dialog";
 export { Timezone } from "./timezone";
 export { SelectedCalendarsSettingsWebWrapper } from "./selected-calendars/wrappers/SelectedCalendarsSettingsWebWrapper";
